Undo prior change, but leave agent_ID UTIL
As Jonathan pointed out, it's not as simple as that to make this compatible with asymmetric games. In particular, the solution in the last commit had the flaw that some agents (agent_0) would always play as males, but others (agent_num_agents) would always play as females, whereas other would sometimes play as males and sometimes play as females. Which role they played depended on the relative position of their and their opponents IDs. This is no good because IDs are not part of the agents' observations. So it's not possible for agents to develop an optimal policy that is a function only of the history of games with their next opponent. For now, we assume that games are symmetric in the way that Prisoner's Dilemma, Stag Hunt, and Chicken are symmetric. Note we can still play games like Battle of the Sexes by using the game {("C", "C"): (0, 0), ("C", "D"): (2, 1), ("D", "C"): (1, 2), ("D", "D"): (0, 0)} and interpreting the "C" as "dominate" and "D" as "submit."